feat(currying): support currying such makeCreator
